Added rich styling into base editor
