+++ src/lib/editor/BaseEditor.js
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { Editor, EditorState } from "draft-js";
+import { RichUtils, Editor, EditorState } from "draft-js";
 import { mutate } from "@react-mutate/core";
 import styled from "styled-components";
 
@@ -15,6 +15,7 @@ const EditorStyles = styled.div`
 
 /**
  * A really basic editor that can type and focus and maintain editor state.
+ * Also can do rich-styling.
  */
 class BaseEditor extends React.Component {
   constructor(props) {
@@ -44,12 +45,21 @@ class BaseEditor extends React.Component {
     if (this.props.onChangeEx) {
       this.props.onChangeEx(editorState);
     }
-    console.log("Change");
   };
 
   componentDidMount() {
     this.handleFocus();
   }
+
+  // rich styling here
+  handleKeyCommand = (command, editorState) => {
+    const newState = RichUtils.handleKeyCommand(editorState, command);
+    if (newState) {
+      this.onChange(newState);
+      return "handled";
+    }
+    return "not-handled";
+  };
 
   render() {
     return (
@@ -59,6 +69,7 @@ class BaseEditor extends React.Component {
           ref={this.setDomEditorRef}
           onChange={this.onChange}
           editorState={this.state.editorState}
+          handleKeyCommand={this.handleKeyCommand}
           {...this.props}
         />
       </EditorStyles>
