--- src/lib/styled-editor/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "styled-editor",
-  "version": "1.0.0",
-  "description": "A basic editor that can support rich styling (bold, italics, etc...)",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC"
-}
